Fix the behavior of ignore_invalid_options
The current implementation of ignore_invalid_options does not include
the ignored options in the `leftovers`. This is a consequence of zero
being truthy in Ruby (in `unless num_params_taken`) and an off-by-one
error in the code for handling that case. This PR fixes `ignore_invalid_options`
and adds a test suite for it.