[ParserTest] Test boolean short form
In 2.0, the logic for parsing boolean flags was changed/standardized:

  https://github.com/ManageIQ/optimist/commit/4ee290fe5709425acb24519fe3e4ea96857e231b

When given a short form, it will take the logic of the "positive"
(non-no) long flag.  This this is the case whether or not the .opt is
defined starting with "no_..." (which has special logic to instead have
the --no-flag option be set to true) or with as a normal flag.

This means that given:

  opt :foo, "Give it foo"

The parsing logic would be that sets :foo would be:

  -f       => true
  --foo    => true
  --no-foo => false

And the following:

  opt :no_foo, "No foos given",

Would map to:

  -n       => false
  --foo    => false
  --no-foo => true